0.0.8 Automatic Looping (InProg)
Trying to make looping automatic
But all I did was put keybinds and stuff. . . .

What I wanna do:
When scene starts
{
Record current run as a Loop
}
When player dies
{
Restart Scene
Play Recorded Loop
}

Bu when restarting the scene it starts recording again, I need to find a way to tell the computer to not record all the time when the scene is played!