Attacks Sequenced | Alpha-0.1.2
I can now scedule settings for the turret to use.

I can set when settings apply and the settings being the cone angles of the attack, speed of bullets and etc.

It's buggy and I'll still need to implement more settings to be tampered with so I can make better attacks